VERSION 0.1 BUILD 50
The Display interface of the console module is changed to private. The console module input interface provides an exit function. The State module provides an interface for switches to automatically adjust coordinates. The drawing module provides an interface for toggling and off to enable adjustments on the move. The drawing module is modified to singleton mode. The engine module provides mobile comprehensive function settings.